Use unsigned bytes instead of signed bytes (tests failing)

+++ Com.Adobe.Xmp/Com/adobe/xmp/impl/Base64.cs
@@ -24,25 +24,25 @@ namespace Com.Adobe.Xmp.Impl
     public static class Base64
     {
         /// <summary>marker for invalid bytes</summary>
-        private const sbyte Invalid = unchecked(-1);
+        private const byte Invalid = unchecked((byte)-1);
 
         /// <summary>marker for accepted whitespace bytes</summary>
-        private const sbyte Whitespace = unchecked(-2);
+        private const byte Whitespace = unchecked((byte)-2);
 
         /// <summary>marker for an equal symbol</summary>
-        private const sbyte Equal = unchecked(-3);
+        private const byte Equal = unchecked((byte)-3);
 
-        private static readonly sbyte[] base64 = new sbyte[] { unchecked((sbyte)(byte)('A')), unchecked((sbyte)(byte)('B')), unchecked((sbyte)(byte)('C')), unchecked((sbyte)(byte)('D')), unchecked((sbyte)(byte)('E')), unchecked((sbyte)(byte)('F')), unchecked(
-            (sbyte)(byte)('G')), unchecked((sbyte)(byte)('H')), unchecked((sbyte)(byte)('I')), unchecked((sbyte)(byte)('J')), unchecked((sbyte)(byte)('K')), unchecked((sbyte)(byte)('L')), unchecked((sbyte)(byte)('M')), unchecked((sbyte)(byte)('N')), unchecked(
-            (sbyte)(byte)('O')), unchecked((sbyte)(byte)('P')), unchecked((sbyte)(byte)('Q')), unchecked((sbyte)(byte)('R')), unchecked((sbyte)(byte)('S')), unchecked((sbyte)(byte)('T')), unchecked((sbyte)(byte)('U')), unchecked((sbyte)(byte)('V')), unchecked(
-            (sbyte)(byte)('W')), unchecked((sbyte)(byte)('X')), unchecked((sbyte)(byte)('Y')), unchecked((sbyte)(byte)('Z')), unchecked((sbyte)(byte)('a')), unchecked((sbyte)(byte)('b')), unchecked((sbyte)(byte)('c')), unchecked((sbyte)(byte)('d')), unchecked(
-            (sbyte)(byte)('e')), unchecked((sbyte)(byte)('f')), unchecked((sbyte)(byte)('g')), unchecked((sbyte)(byte)('h')), unchecked((sbyte)(byte)('i')), unchecked((sbyte)(byte)('j')), unchecked((sbyte)(byte)('k')), unchecked((sbyte)(byte)('l')), unchecked(
-            (sbyte)(byte)('m')), unchecked((sbyte)(byte)('n')), unchecked((sbyte)(byte)('o')), unchecked((sbyte)(byte)('p')), unchecked((sbyte)(byte)('q')), unchecked((sbyte)(byte)('r')), unchecked((sbyte)(byte)('s')), unchecked((sbyte)(byte)('t')), unchecked(
-            (sbyte)(byte)('u')), unchecked((sbyte)(byte)('v')), unchecked((sbyte)(byte)('w')), unchecked((sbyte)(byte)('x')), unchecked((sbyte)(byte)('y')), unchecked((sbyte)(byte)('z')), unchecked((sbyte)(byte)('0')), unchecked((sbyte)(byte)('1')), unchecked(
-            (sbyte)(byte)('2')), unchecked((sbyte)(byte)('3')), unchecked((sbyte)(byte)('4')), unchecked((sbyte)(byte)('5')), unchecked((sbyte)(byte)('6')), unchecked((sbyte)(byte)('7')), unchecked((sbyte)(byte)('8')), unchecked((sbyte)(byte)('9')), unchecked(
-            (sbyte)(byte)('+')), unchecked((sbyte)(byte)('/')) };
+        private static readonly byte[] base64 = new byte[] { unchecked((byte)(byte)('A')), unchecked((byte)(byte)('B')), unchecked((byte)(byte)('C')), unchecked((byte)(byte)('D')), unchecked((byte)(byte)('E')), unchecked((byte)(byte)('F')), unchecked(
+            (byte)(byte)('G')), unchecked((byte)(byte)('H')), unchecked((byte)(byte)('I')), unchecked((byte)(byte)('J')), unchecked((byte)(byte)('K')), unchecked((byte)(byte)('L')), unchecked((byte)(byte)('M')), unchecked((byte)(byte)('N')), unchecked(
+            (byte)(byte)('O')), unchecked((byte)(byte)('P')), unchecked((byte)(byte)('Q')), unchecked((byte)(byte)('R')), unchecked((byte)(byte)('S')), unchecked((byte)(byte)('T')), unchecked((byte)(byte)('U')), unchecked((byte)(byte)('V')), unchecked(
+            (byte)(byte)('W')), unchecked((byte)(byte)('X')), unchecked((byte)(byte)('Y')), unchecked((byte)(byte)('Z')), unchecked((byte)(byte)('a')), unchecked((byte)(byte)('b')), unchecked((byte)(byte)('c')), unchecked((byte)(byte)('d')), unchecked(
+            (byte)(byte)('e')), unchecked((byte)(byte)('f')), unchecked((byte)(byte)('g')), unchecked((byte)(byte)('h')), unchecked((byte)(byte)('i')), unchecked((byte)(byte)('j')), unchecked((byte)(byte)('k')), unchecked((byte)(byte)('l')), unchecked(
+            (byte)(byte)('m')), unchecked((byte)(byte)('n')), unchecked((byte)(byte)('o')), unchecked((byte)(byte)('p')), unchecked((byte)(byte)('q')), unchecked((byte)(byte)('r')), unchecked((byte)(byte)('s')), unchecked((byte)(byte)('t')), unchecked(
+            (byte)(byte)('u')), unchecked((byte)(byte)('v')), unchecked((byte)(byte)('w')), unchecked((byte)(byte)('x')), unchecked((byte)(byte)('y')), unchecked((byte)(byte)('z')), unchecked((byte)(byte)('0')), unchecked((byte)(byte)('1')), unchecked(
+            (byte)(byte)('2')), unchecked((byte)(byte)('3')), unchecked((byte)(byte)('4')), unchecked((byte)(byte)('5')), unchecked((byte)(byte)('6')), unchecked((byte)(byte)('7')), unchecked((byte)(byte)('8')), unchecked((byte)(byte)('9')), unchecked(
+            (byte)(byte)('+')), unchecked((byte)(byte)('/')) };
 
-        private static readonly sbyte[] Ascii = new sbyte[255];
+        private static readonly byte[] Ascii = new byte[255];
 
         static Base64()
         {
@@ -70,7 +70,7 @@ namespace Com.Adobe.Xmp.Impl
             // valid bytes
             for (int idx1 = 0; idx1 < base64.Length; idx1++)
             {
-                Ascii[base64[idx1]] = unchecked((sbyte)idx1);
+                Ascii[base64[idx1]] = unchecked((byte)idx1);
             }
             // whitespaces
             Ascii[unchecked(0x09)] = Whitespace;
@@ -88,7 +88,7 @@ namespace Com.Adobe.Xmp.Impl
         /// must be dividable by four; 0 means no linefeeds
         /// </param>
         /// <returns>the base64-encoded data.</returns>
-        public static sbyte[] Encode(sbyte[] src, int lineFeed = 0)
+        public static byte[] Encode(byte[] src, int lineFeed = 0)
         {
             // linefeed must be dividable by 4
             lineFeed = lineFeed / 4 * 4;
@@ -102,7 +102,7 @@ namespace Com.Adobe.Xmp.Impl
             {
                 codeLength += (codeLength - 1) / lineFeed;
             }
-            sbyte[] dst = new sbyte[codeLength];
+            byte[] dst = new byte[codeLength];
             int bits24;
             int bits6;
             //
@@ -140,7 +140,7 @@ namespace Com.Adobe.Xmp.Impl
                 dst[didx++] = base64[bits6];
                 bits6 = (bits24 & unchecked(0x00000FC0)) >> 6;
                 dst[didx++] = base64[bits6];
-                dst[didx++] = unchecked((sbyte)(byte)('='));
+                dst[didx++] = unchecked((byte)(byte)('='));
             }
             else
             {
@@ -151,8 +151,8 @@ namespace Com.Adobe.Xmp.Impl
                     dst[didx++] = base64[bits6];
                     bits6 = (bits24 & unchecked(0x0003F000)) >> 12;
                     dst[didx++] = base64[bits6];
-                    dst[didx++] = unchecked((sbyte)(byte)('='));
-                    dst[didx++] = unchecked((sbyte)(byte)('='));
+                    dst[didx++] = unchecked((byte)(byte)('='));
+                    dst[didx++] = unchecked((byte)(byte)('='));
                 }
             }
             return dst;
@@ -173,7 +173,7 @@ namespace Com.Adobe.Xmp.Impl
         /// Thrown if the base 64 strings contains non-valid characters,
         /// beside the bas64 chars, LF, CR, tab and space are accepted.
         /// </exception>
-        public static sbyte[] Decode(sbyte[] src)
+        public static byte[] Decode(byte[] src)
         {
             //
             // Do ascii printable to 0-63 conversion.
@@ -182,7 +182,7 @@ namespace Com.Adobe.Xmp.Impl
             int srcLen = 0;
             for (sidx = 0; sidx < src.Length; sidx++)
             {
-                sbyte val = Ascii[src[sidx]];
+                byte val = Ascii[src[sidx]];
                 if (val >= 0)
                 {
                     src[srcLen++] = val;
@@ -202,24 +202,24 @@ namespace Com.Adobe.Xmp.Impl
             {
                 srcLen--;
             }
-            sbyte[] dst = new sbyte[srcLen * 3 / 4];
+            byte[] dst = new byte[srcLen * 3 / 4];
             //
             // Do 4-byte to 3-byte conversion.
             //
             int didx;
             for (sidx = 0, didx = 0; didx < dst.Length - 2; sidx += 4, didx += 3)
             {
-                dst[didx] = unchecked((sbyte)(((src[sidx] << 2) & unchecked(0xFF)) | ((src[sidx + 1] >> 4) & unchecked(0x03))));
-                dst[didx + 1] = unchecked((sbyte)(((src[sidx + 1] << 4) & unchecked(0xFF)) | ((src[sidx + 2] >> 2) & unchecked(0x0F))));
-                dst[didx + 2] = unchecked((sbyte)(((src[sidx + 2] << 6) & unchecked(0xFF)) | ((src[sidx + 3]) & unchecked(0x3F))));
+                dst[didx] = unchecked((byte)(((src[sidx] << 2) & unchecked(0xFF)) | ((src[sidx + 1] >> 4) & unchecked(0x03))));
+                dst[didx + 1] = unchecked((byte)(((src[sidx + 1] << 4) & unchecked(0xFF)) | ((src[sidx + 2] >> 2) & unchecked(0x0F))));
+                dst[didx + 2] = unchecked((byte)(((src[sidx + 2] << 6) & unchecked(0xFF)) | ((src[sidx + 3]) & unchecked(0x3F))));
             }
             if (didx < dst.Length)
             {
-                dst[didx] = unchecked((sbyte)(((src[sidx] << 2) & unchecked(0xFF)) | ((src[sidx + 1] >> 4) & unchecked(0x03))));
+                dst[didx] = unchecked((byte)(((src[sidx] << 2) & unchecked(0xFF)) | ((src[sidx + 1] >> 4) & unchecked(0x03))));
             }
             if (++didx < dst.Length)
             {
-                dst[didx] = unchecked((sbyte)(((src[sidx + 1] << 4) & unchecked(0xFF)) | ((src[sidx + 2] >> 2) & unchecked(0x0F))));
+                dst[didx] = unchecked((byte)(((src[sidx + 1] << 4) & unchecked(0xFF)) | ((src[sidx + 2] >> 2) & unchecked(0x0F))));
             }
             return dst;
         }
